Extract cluster transport layer into cluster_link.{c,h}
Introduce clusterMsgSendBlock as a generic refcounted send buffer with
uint64_t data[] for alignment. Move clusterLink struct, link flags,
RCVBUF_INIT_LEN, createClusterLink, freeClusterLink,
setClusterNodeToInboundClusterLink, clusterWriteHandler,
clusterMsgSendBlockDecrRefCount, clusterLinkGetNodeName,
clusterLinkGetHumanNodeName, and CLUSTER LINKS reply functions from
cluster_legacy.c to cluster_link.{c,h}.

The write handler now sends block->totlen bytes from block->data,
making it protocol-agnostic. Legacy code uses getMessageFromSendBlock()
to cast the data to clusterMsg pointers.
